wip: split airdrop items to their own files
