+++ src/assets/icons/LinkIcon.tsx
@@ -1,0 +1,11 @@
+const LinkIcon = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
+        <circle cx="12" cy="12" r="12" fill="currentColor" />
+        <path
+            d="M13.4849 11.2279C13.4895 11.3055 13.4895 11.3877 13.4713 11.4653C13.4713 11.4699 13.4667 11.4745 13.4667 11.4745C13.4347 11.6114 13.3662 11.7393 13.2612 11.8489C12.9553 12.1548 12.4576 12.1548 12.1517 11.8489C11.8367 11.5338 11.3938 11.4151 10.9829 11.4927C10.8961 11.511 10.8139 11.5338 10.7317 11.5658C10.5765 11.6297 10.4304 11.7256 10.3025 11.8489L9.07888 13.0771L7.59497 14.561C7.08816 15.0678 7.08816 15.8988 7.59497 16.4056C8.08809 16.8987 8.94647 16.8987 9.43958 16.4056L10.5674 15.2778C11.0422 15.602 11.5764 15.8212 12.1426 15.9171L10.5491 17.5106C9.98749 18.0722 9.25238 18.3507 8.51728 18.3507C7.78217 18.3507 7.04707 18.0722 6.48546 17.5106C5.36682 16.3919 5.36682 14.5656 6.48546 13.4469L9.19303 10.7394C9.75463 10.1778 10.4897 9.89925 11.2248 9.89925C11.9599 9.89925 12.6951 10.1778 13.2567 10.7394C13.316 10.7987 13.3617 10.8672 13.3982 10.9403C13.4028 10.9494 13.4119 10.9631 13.4165 10.9722C13.4484 11.0453 13.4667 11.1183 13.4758 11.1914C13.4849 11.2051 13.4849 11.2188 13.4849 11.2279ZM17.5942 7.19168C17.0509 6.64834 16.3295 6.35156 15.5624 6.35156C14.7954 6.35156 14.0739 6.64834 13.5306 7.19168L11.9326 8.78974C12.4896 8.88562 13.0284 9.10022 13.5078 9.42896L14.6355 8.30119C15.1287 7.80808 15.987 7.80808 16.4802 8.30119C16.7267 8.54775 16.8637 8.87649 16.8637 9.2235C16.8637 9.5705 16.7267 9.89925 16.4802 10.1458L15.9825 10.6435L13.7772 12.8534C13.284 13.3465 12.4257 13.3465 11.9326 12.8534C11.6266 12.5475 11.129 12.5475 10.823 12.8534C10.5171 13.1593 10.5171 13.657 10.823 13.9629C11.3664 14.5062 12.0878 14.803 12.8549 14.803C13.6219 14.803 14.3433 14.5062 14.8867 13.9629L17.0966 11.753L17.5942 11.2553C18.1376 10.712 18.4344 9.99056 18.4344 9.2235C18.4344 8.45643 18.1376 7.73502 17.5942 7.19168Z"
+            fill="currentColor"
+        />
+    </svg>
+);
+
+export default LinkIcon;
