+++ src/containers/main/Airdrop/sidebar/AirdropSidebarItems.tsx
@@ -1,0 +1,14 @@
+import Invite from './Invite.tsx';
+import Gems from './Gems.tsx';
+import User from './User.tsx';
+import { Wrapper } from './items.style';
+
+export function AirdropSidebarItems() {
+    return (
+        <Wrapper>
+            <Invite />
+            <Gems />
+            <User />
+        </Wrapper>
+    );
+}
