+++ src/containers/main/Airdrop/sidebar/components/item.style.ts
